+++ fdm-app/app/components/blocks/field-dashboard/registry.tsx
@@ -78,7 +78,6 @@ export function buildFieldDashboardRegistry({
       section: "anchor",
       title: "Perceeldetails",
       span: "xl:col-span-4",
-      permissionGate: "field-write",
       detailHref: "./settings",
       dataKey: "field",
       Component: FieldDashboardIdentityTile,
@@ -88,7 +87,6 @@ export function buildFieldDashboardRegistry({
       section: "gewas",
       title: "Hoofdgewas",
       span: "xl:col-span-7",
-      permissionGate: "field-write",
       detailHref: "./cultivation",
       dataKey: "cultivation.active",
       Component: FieldDashboardCurrentCultivationTile,
@@ -107,7 +105,6 @@ export function buildFieldDashboardRegistry({
       section: "bemesting",
       title: "Bemestingen",
       span: "",
-      permissionGate: "field-write",
       detailHref: "./fertilizer",
       dataKey: "fertilizer",
       Component: FieldDashboardFertilizerApplicationsTile,
@@ -189,7 +186,6 @@ export function buildFieldDashboardRegistry({
       section: "bodemindicatoren",
       title: "Maatregelen",
       span: "xl:col-span-1",
-      permissionGate: "field-write",
       detailHref: fieldMeasuresHref,
       dataKey: "measures",
       Component: FieldDashboardMeasuresTile,
