+++ fdm-app/app/routes/farm.$b_id_farm.$calendar.field.$b_id._index.tsx
@@ -770,7 +770,11 @@ export default function FieldDashboardRoute() {
         if (tiles.length === 0) return null
 
         return (
-          <section key={section.id} className="space-y-4" aria-labelledby={`${section.id}-heading`}>
+          <section
+            key={section.id}
+            className="space-y-4"
+            aria-labelledby={section.showHeading !== false ? `${section.id}-heading` : undefined}
+          >
             {section.showHeading !== false ? (
               <div id={`${section.id}-heading`}>
                 <FieldDashboardSectionHeading title={section.title} />
